+++ src/providers/comscore/comscore.js
@@ -34,3 +34,4 @@ analytics.addProvider('comScore', {
 
 });
 
+