+++ src/providers/gauges/gauges.js
@@ -40,3 +40,5 @@ analytics.addProvider('Gauges', {
     }
 
 });
+
+